move armor patterns to registry package
